Feat: 업적 평가 항목 작성중 수정
- 업적 평가 항목 작성중 상태로 수정

+++ grid/src/main/java/org/highfives/grid/performance_review/command/dto/PerformanceReviewItemDTO.java
@@ -17,9 +17,9 @@ public class PerformanceReviewItemDTO {
     private int weight;
     private String performance;
     private int selfId;
-    private float selfScore;
+    private double selfScore;
     private String selfComment;
     private int superiorId;
-    private float superiorScore;
+    private double superiorScore;
     private int reviewId;
 }
